separated inbox and basic disk logic.
